% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versions.R
\name{dryad_version}
\alias{dryad_version}
\title{Get metadata for a single dataset version}
\usage{
dryad_version(version_id)
}
\arguments{
\item{version_id}{Numeric or character version id, as reported in
\code{\link[=dryad_dataset_versions]{dryad_dataset_versions()}}.}
}
\value{
A nested list with the version metadata.
}
\description{
Wraps \code{GET /versions/{id}}.
}
\examples{
\dontrun{
dryad_use_sandbox()
dryad_version(26724)
}
}
